Add files via upload
Database, which will store all the links,  names, dates, coments and etc for the bot
Currently have all 3 kinds of entertainments in the corresponding list. Also contains lists of names, unique and non-unique IDs and links (only in film_list) in the corresponding lists. In the future at least 2 lists will be added. 

Next stage is connecting db to the telebot